KP-8800 RosterVector should consist of int`s

+++ src/Tests/WB.Tests.Unit/SharedKernels/Enumerator/StatefulInterviewTests/when_looking_for_options_of_linked_question_which_linked_on_nested_roster_question.cs
@@ -55,13 +55,13 @@ namespace WB.Tests.Unit.SharedKernels.Enumerator.StatefulInterviewTests
         It should_order_options_by_roster_sort_index_at_first = () =>
         {
             interview.GetLinkedMultiOptionQuestion(linkedToQuestionIdentity)
-            .Options.Select(a => a.First()).ToArray().ShouldEqual(new decimal[] {1, 1, 2, 2});
+            .Options.Select(a => a.First()).ToArray().ShouldEqual(new int[] {1, 1, 2, 2});
         };
 
         It should_order_options_by_nested_roster_sort_index_in_scope_of_parent_roster = () =>
         {
             interview.GetLinkedMultiOptionQuestion(linkedToQuestionIdentity)
-            .Options.Select(a => a.Last()).ToArray().ShouldEqual(new decimal[] { 1, 2, 1, 2 });
+            .Options.Select(a => a.Last()).ToArray().ShouldEqual(new int[] { 1, 2, 1, 2 });
         };
 
         private static StatefulInterview interview;
